+++ async-race/src/app/view/pages/garage/Garage.ts
@@ -2,7 +2,7 @@ import API from '../../../../api/Api';
 import { EErrors } from '../../../../typescript/enums';
 import { TColorHEX } from '../../../../typescript/types';
 import PageBuilder from '../../../utils/PageBuilder';
-import ErrorView from '../components/errorView/ErrorView';
+import ErrorView from '../components/errorView/errorView';
 import Editor from './editor/Editor';
 
 import GarageTable from './table/GarageTable';
